feat: HomeFeature에서 map 관련 분리
