+++ Projects/Sseudam/Sources/DependencyRegister.swift
@@ -40,7 +40,6 @@ struct DependencyRegister {
   /// UseCase에 Repository 주입
   func injection() {
     let networker = NetworkKit()
-    let homeRepository = HomeRepository.live
     let authRepository = AuthRepository.live(networker: networker)
     let userReoository = UserRepository.live(networker: networker)
     let trashSpotRepository = TrashSpotRepository.live(networker: networker)
